# route_schedule: Improve ranked pairs implementation