http://jira.canaltp.fr/browse/NAVITIAII-1988

Bench for RER and Transilien on fr-idf (kraken only):

Line | removed | topo before | time before | topo after | time after
------|-------------|----------------|----------------|--------------|-------------
A | 1314 | 149644 | 24.1 | 19601 | 6.0
B |   692 | 105984 | 14.1 |   9939 | 2.4
C |   364 |   99623 | 11.8 |   5308 | 1.5
D | 1453 |   94221 | 10.3 | 21426 | 4.8
E |   677 |   77381 |   8.5 |   9570 | 1.8
H | 1849 |   98187 | 10.8 | 28425 | 4.8
J |    712 | 107003 | 13.5 | 11383 | 2.3
K |       0 |           1 |   0.0 |         1 | 0.0
L |   159 | 179827 | 33.0 |   2555 | 1.0
N |     47 |   57244 |   4.2 |     672 | 0.2
P |     46 |   27357 |   1.9 |     638 | 0.1
R |     36 |     6653 |   0.2 |     405 | 0.0
U |       0 |           1 |   0.0 |         1 | 0.0

nb edges = topo before - 2 (except when topo before = 1)

tldr: always faster, really faster on L