@TeXitoi 2016-01-14
@l-vincent-l small improvement of the thermometer heuristic!