@TeXitoi 2016-01-14
For Transilien L on fr-idf, before 32s, after 2s (+ 5s of jormungandr).